requirements.txt
django 
djangorestframework 
pillow 
setuptools 
django-cors-headers
drf-yasg 
coreapi